Separate TwoFactor into new trait rather than SessionGuard
